Release of version 3.1.0
CADET has been extended to support networks of unit operations. Many
more changes have been made and the most important ones are listed
below. Note that the interface has changed and breaking changes had to
be introduced.

Major changes:
- Full support for networks of unit operations (CADET now natively
simulates SMB and other multi-column processes)
- Support for backwards flow in General Rate Model
- Replace OpenMP parallelization by Intel Threading Building Blocks
- Add some behavior tests using the CATCH unit testing framework
- Extended documentation (improved manual)
- Improved CMake build system

+++ include/cadet/ParameterId.hpp
@@ -1,7 +1,7 @@
 // =============================================================================
 //  CADET - The Chromatography Analysis and Design Toolkit
 //  
-//  Copyright © 2008-2016: The CADET Authors
+//  Copyright © 2008-2017: The CADET Authors
 //            Please see the AUTHORS and CONTRIBUTORS file.
 //  
 //  All rights reserved. This program and the accompanying materials
@@ -29,7 +29,7 @@
 namespace cadet
 {
 
-typedef uint8_t UnitOpIdx;
+typedef uint16_t UnitOpIdx;
 typedef uint8_t ComponentIdx;
 typedef uint8_t BoundPhaseIdx;
 typedef uint8_t ReactionIdx;
